requirements.txt
pillow
torchmetrics
pandas